requirements_web.txt
Flask==2.3.3
yt-dlp==2025.9.26
pytube==15.0.0
requests==2.31.0
BeautifulSoup4==4.12.2
qrcode==7.4.2
stripe==6.0.0
paypalrestsdk==1.13.1
Werkzeug==2.3.7
Flask-WTF==1.1.1
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.20
WTForms==3.0.1
gunicorn==21.2.0
python-dotenv==1.0.0
python-slugify==8.0.4
Authlib==1.3.0
Flask-Mail==0.9.1